requirements.txt
pytelegrambotapi
ujson
websockets
aiohttp
pyyaml
atlassian-python-api
psycopg2-binary
requests

